Middleware: Moved Controller class namespace
