+++ tests/Controllers/AccessDenied.php
@@ -1,6 +1,8 @@
 <?php
 
-class AccessDenied extends \Docnet\JAPI\Controller
+use Docnet\JAPI\Controller\Controller;
+
+class AccessDenied extends Controller
 {
     public function dispatch(){
         throw new \Docnet\JAPI\Exceptions\AccessDenied('Error Message', 403);
