+++ tests/Controllers/Example.php
@@ -1,9 +1,10 @@
 <?php
 
+use Docnet\JAPI\Controller\Controller;
 use gordonmcvey\httpsupport\enum\statuscodes\SuccessCodes;
 use gordonmcvey\httpsupport\Response;
 
-class Example extends \Docnet\JAPI\Controller
+class Example extends Controller
 {
     public function dispatch(){
         $this->setResponse(new Response(SuccessCodes::OK, json_encode(['test' => true])));
