+++ tests/Controllers/Exceptional.php
@@ -1,6 +1,8 @@
 <?php
 
-class Exceptional extends \Docnet\JAPI\Controller
+use Docnet\JAPI\Controller\Controller;
+
+class Exceptional extends Controller
 {
     public function dispatch(){
         throw new RuntimeException('Error Message', 400);
