+++ tests/Controllers/Whoops.php
@@ -1,6 +1,8 @@
 <?php
 
-class Whoops extends \Docnet\JAPI\Controller
+use Docnet\JAPI\Controller\Controller;
+
+class Whoops extends Controller
 {
     public function dispatch(){
         throw new Exception;
